add some checkboxes and backend renaming
