+++ routes/favorites.js
